Add ability to ES sync specify models
This adds a new argument to the `sync_es` command.

When `--model=<name>` is used, only the model <name> is synced.
You can sync multiple models at once with
`--model=<model1> --model=<model2>` etc.

If no --model is specified, all models are synced as usual.
